# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Margen",
    "description": "Asistente financiero personal. Todos los montos viajan como entero de centavos; el servidor no emite decimales en ningún camino monetario.",
    "version": "v1"
  },
  "paths": {
...
    "/auth/devices": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/challenges": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/tokens": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Autenticación"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/auth/shortcut-tokens": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortcutTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/tokens/{tokenId}": {
      "delete": {
        "tags": [
          "Autenticación"
        ],
        "parameters": [
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/budget/redistribute": {
      "post": {
        "tags": [
          "Presupuesto"
        ],
        "operationId": "RedistributeBudget",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedistributeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChallengeRequest": {
        "required": [
          "deviceId"
        ],
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "RedeemRequest": {
        "required": [
          "deviceId",
          "nonce",
          "signature"
        ],
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "nonce": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          }
        }
      },
      "RedistributeRequest": {
        "required": [
          "fromCategoryId",
          "toCategoryId",
          "amountCents"
        ],
        "type": "object",
        "properties": {
          "fromCategoryId": {
            "type": "string",
            "format": "uuid"
          },
          "toCategoryId": {
            "type": "string",
            "format": "uuid"
          },
          "amountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "RegisterDeviceRequest": {
        "required": [
          "name",
          "publicKeySpki"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "publicKeySpki": {
            "type": "string"
          }
        }
      },
      "ShortcutTokenRequest": {
        "required": [
          "label"
        ],
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          }
        }
      }
    }
  },
...
}